server/requirements.txt
fastapi
sqlmodel
uvicorn
mysql-connector-python
python-dotenv
websockets
pygame
pymysql